requirements.txt
backcall==0.2.0
bert-score==0.3.6
block==0.0.5
block.bootstrap.pytorch==0.1.5
bootstrap.pytorch==0.0.13
certifi==2020.12.5
click==8.0.0
cycler==0.10.0
decorator==4.4.2
future==0.18.2
h5py==2.7.1
imageio==2.9.0
ipdb==0.13.7
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.18.0
kiwisolver==1.3.1
matplotlib==3.3.4
munch==2.5.0
networkx==2.5.1
nltk==3.3
numpy==1.19.5
opencv-python==4.5.2.52
pandas==1.1.2
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
plotly==4.14.3
pretrainedmodels==0.7.4
prompt-toolkit==3.0.18
protobuf==3.17.0
ptyprocess==0.7.0
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
retrying==1.3.3
scikit-image==0.17.2
scipy==1.5.4
seaborn==0.11.1
six==1.16.0
skipthoughts==0.0.1
tabulate==0.8.9
tensorboardX==2.2
tifffile==2020.9.3
toml==0.10.2
#torch==1.6.0
#torchvision==0.7.0
tqdm==4.60.0
traitlets==4.3.3
wcwidth==0.2.5
